update version to 0.47.0-public-SNAPSHOT

+++ odps-sdk/odps-sdk-core/src/main/java/com/aliyun/odps/tunnel/impl/UpsertRecord.java
@@ -63,7 +63,7 @@ public class UpsertRecord extends ArrayRecord {
         record.setTinyint(TunnelConstants.META_FIELD_OPERATION, operation);
     }
 
-    void setValueCols(List<Integer> cols) {
+    void setValueCols(List<Long> cols) {
         record.setArray(TunnelConstants.META_FIELD_VALUE_COLS, cols);
     }
